Giving meaning to the Volume struct Add impl Default (to 50% = 0x8000) Move the conversion to logarithmic volume to Volumes impl.
